Moved the variable ACE_TAO_OPTION to the packageconfig 'doc-group3'
The selection to use ACE_TAO version 2 or 3 is is now in the
packageconfig 'doc-group3'.
Default the version 2 is used.  As it is in the configuration of
OpenDDS.

Although the syntax was correct before,  using variables is tainting
the global namespace of Yocto.
And using a packageconfig is more the yocto way.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>