opendds_git: rework of the Git version of the recipe
The previous version of the OpenDDS recipe, which was meant to work with the latest status of the Git repository, was no longer functional. This commit fixes the issues.

Currently, only the OpenDDS Git repository is at the master branch. The ACE/TAO and MPC components are, for now, set to their latest releases.

Additionally, the download process has been shifted from the configure script to the Yocto recipe.

If the configure script cannot find ACE, TAO, and MPC, it will attempt to download them. However, this is not ideal for Yocto. Downloads should occur in the do_fetch step for the following reasons:

    To ensure reproducible builds.
    To support air-gapped networks, which might be used by OpenDDS users.

TODO:
  TAO, ACE and MPC are still the releases as defined in ace_tao.ini and not the latest version of git

Tested with ptest and ishapes demo at Raspberry Pi 4.

Signed-off-by: Jan Vermaete <jan.vermaete@gmail.com>